+++ client/src/App.js
@@ -1,3 +1,4 @@
+import {DEFAULT_THEME} from "AppAppearance";
 import axios from "axios";
 import ComponentLoader from "components/ComponentLoader";
 import AppContext from "context/AppContext";
@@ -9,28 +10,28 @@ import {toast} from "react-toastify";
 
 import ErrorRoute from "routes/ErrorRoute";
 import LoadingRouteUtil from "routes/utils/LoadingRouteUtil";
-import tinycolor from "tinycolor2";
 import {getCookieOrDefault} from "utils/basic-utils";
 import {retry} from "utils/lazy-init";
 
-const ProfileView = lazy(() => retry(() => import("routes/profile/ProfileRoute")));
-const CreateBoardView = lazy(() => retry(() => import("routes/board/creator/CreatorBoardRoute")));
-const ModeratorInvitation = lazy(() => retry(() => import("routes/ModeratorInvitationRoute")));
-const BoardView = lazy(() => retry(() => import("routes/board/BoardRoute")));
-const RoadmapView = lazy(() => retry(() => import("routes/RoadmapRoute")));
-const AdminPanelView = lazy(() => retry(() => import("routes/board/admin/BoardAdminPanelRoute")));
-const IdeaView = lazy(() => retry(() => import("routes/IdeaRoute")));
-const OauthReceiver = lazy(() => retry(() => import("routes/LoginRoute")));
-const UnsubscribeView = lazy(() => retry(() => import("routes/NotificationUnsubscribeRoute")));
+const ProfileRoute = lazy(() => retry(() => import("routes/profile/ProfileRoute")));
+const CreateBoardRoute = lazy(() => retry(() => import("routes/board/creator/CreatorBoardRoute")));
+const ModeratorInvitationRoute = lazy(() => retry(() => import("routes/ModeratorInvitationRoute")));
+const BoardRoute = lazy(() => retry(() => import("routes/board/BoardRoute")));
+const RoadmapRoute = lazy(() => retry(() => import("routes/RoadmapRoute")));
+const BoardAdminPanelRoute = lazy(() => retry(() => import("routes/board/admin/BoardAdminPanelRoute")));
+const IdeaRoute = lazy(() => retry(() => import("routes/IdeaRoute")));
+const LoginRoute = lazy(() => retry(() => import("routes/LoginRoute")));
+const NotificationUnsubscribeRoute = lazy(() => retry(() => import("routes/NotificationUnsubscribeRoute")));
 const UiTestRoute = lazy(() => retry(() => import("routes/UiTestRoute")));
 
 toast.configure();
 
 const CLIENT_VERSION = "1.0.0-alpha";
 const API_ROUTE = (process.env.REACT_APP_SERVER_IP_ADDRESS || "https://app.feedbacky.net") + "/api/v1";
-const DEFAULT_THEME = "#343a40";
 
-const App = () => {
+const App = ({appearanceSettings}) => {
+    console.log(appearanceSettings);
+    const {theme, setTheme, darkMode, getTheme, onDarkModeToggle} = appearanceSettings;
     const [session, setSession] = useState(Cookies.get("FSID"));
     const [localPrefs, setLocalPrefs] = useState({
         ideas: {filter: getCookieOrDefault("prefs_searchFilter", ""), sort: getCookieOrDefault("prefs_searchSort", "")},
@@ -38,21 +39,10 @@ const App = () => {
     });
     const [serviceData, setServiceData] = useState({loaded: false, data: [], error: false});
     const [userData, setUserData] = useState({loaded: false, data: [], error: false});
-    const [darkMode, setDarkMode] = useState(getCookieOrDefault("prefs_darkMode", "false") === 'true');
-    const [theme, setTheme] = useState(DEFAULT_THEME);
-
     useEffect(() => {
-        if (darkMode) {
-            document.body.classList.add("dark");
-        } else if (getCookieOrDefault("prefs_darkMode", null) == null && window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches) {
-            setDarkMode(true);
-            document.body.classList.add("dark");
-            Cookies.set("prefs_darkMode", "true", {expires: 10 * 365 * 7 /* 10 years */});
-        }
         axios.defaults.baseURL = API_ROUTE;
         axios.defaults.headers.common["Authorization"] = "Bearer " + session;
-        // eslint-disable-next-line
-    }, []);
+    });
     useEffect(() => {
         if (serviceData.loaded) {
             return;
@@ -114,32 +104,10 @@ const App = () => {
         }
         setLocalPrefs(data);
     };
-    const getTheme = (adjustColor = true) => {
-        let color = tinycolor(theme);
-        if (darkMode && adjustColor) {
-            color = color.lighten(10);
-            //if still not readable, increase again
-            if (tinycolor.readability(color, "#282828") < 2.5) {
-                color = color.lighten(25);
-            }
-        }
-        return color.clone();
-    };
-    const onDarkModeToggle = () => {
-        let darkModeEnabled = (getCookieOrDefault("prefs_darkMode", "false") === 'true');
-        Cookies.set("prefs_darkMode", (!darkModeEnabled).toString(), {expires: 10 * 365 * 7 /* 10 years */});
-        if (darkMode) {
-            document.body.classList.remove("dark");
-        } else {
-            document.body.classList.add("dark");
-        }
-        setDarkMode(!darkMode);
-    };
     if (serviceData.error) {
         return <BrowserRouter><ErrorRoute Icon={FaDizzy} message={"Service Is Temporarily Unavailable"}/></BrowserRouter>
     }
-    return <ComponentLoader loader={<LoadingRouteUtil/>}
-                            loaded={serviceData.loaded && userData.loaded} component={
+    return <ComponentLoader loader={<LoadingRouteUtil/>} loaded={serviceData.loaded && userData.loaded} component={
         <AppContext.Provider value={{
             user: {
                 data: userData.data,
@@ -161,17 +129,17 @@ const App = () => {
         }}>
             <Suspense fallback={<LoadingRouteUtil/>}>
                 <Switch>
-                    <Route exact path={"/"} component={ProfileView}/>
-                    <Route exact path={"/admin/create"} component={CreateBoardView}/>
-                    <Route path={"/me/:section"} component={ProfileView}/>
-                    <Route path={"/me/"} component={ProfileView}/>
-                    <Route path={"/moderator_invitation/:code"} component={ModeratorInvitation}/>
-                    <Route path={"/b/:id/roadmap"} component={RoadmapView}/>
-                    <Route path={"/b/:id"} component={BoardView}/>
-                    <Route path={"/ba/:id"} component={AdminPanelView}/>
-                    <Route path={"/i/:id"} component={IdeaView}/>
-                    <Route path={"/unsubscribe/:id/:code"} component={UnsubscribeView}/>
-                    <Route path={"/auth/:provider"} render={() => <OauthReceiver onLogin={onLogin}/>}/>
+                    <Route exact path={"/"} component={ProfileRoute}/>
+                    <Route exact path={"/admin/create"} component={CreateBoardRoute}/>
+                    <Route path={"/me/:section"} component={ProfileRoute}/>
+                    <Route path={"/me/"} component={ProfileRoute}/>
+                    <Route path={"/moderator_invitation/:code"} component={ModeratorInvitationRoute}/>
+                    <Route path={"/b/:id/roadmap"} component={RoadmapRoute}/>
+                    <Route path={"/b/:id"} component={BoardRoute}/>
+                    <Route path={"/ba/:id"} component={BoardAdminPanelRoute}/>
+                    <Route path={"/i/:id"} component={IdeaRoute}/>
+                    <Route path={"/unsubscribe/:id/:code"} component={NotificationUnsubscribeRoute}/>
+                    <Route path={"/auth/:provider"} render={() => <LoginRoute onLogin={onLogin}/>}/>
                     <Route path={"/-/ui-test"} component={UiTestRoute}/>
                     <Route render={() => <ErrorRoute Icon={FaExclamationCircle} message={"Content Not Found"}/>}/>
                 </Switch>